Separate tests for check_schema() and ModelConfig.from_dict().
