+++ tests/ludwig/config_validation/test_validate_config_combiner.py
@@ -1,9 +1,8 @@
 import pytest
 
-from ludwig.config_validation.validation import get_schema
+from ludwig.config_validation.validation import check_schema, get_schema
 from ludwig.constants import MODEL_ECD, TRAINER
 from ludwig.error import ConfigValidationError
-from ludwig.schema.model_types.base import ModelConfig
 from tests.integration_tests.utils import binary_feature, category_feature, number_feature
 
 
@@ -39,15 +38,17 @@ def test_config_tabnet(eval_batch_size):
             "early_stop": 20,
             "learning_rate": 0.02,
             "optimizer": {"type": "adam"},
-            "decay": True,
-            "decay_steps": 20000,
-            "decay_rate": 0.9,
-            "staircase": True,
+            "learning_rate_scheduler": {
+                "decay": "linear",
+                "decay_steps": 20000,
+                "decay_rate": 0.9,
+                "staircase": True,
+            },
             "regularization_lambda": 1,
             "regularization_type": "l2",
         },
     }
-    ModelConfig.from_dict(config)
+    check_schema(config)
 
 
 def test_config_bad_combiner():
@@ -63,22 +64,22 @@ def test_config_bad_combiner():
     }
 
     # config is valid at this point
-    ModelConfig.from_dict(config)
+    check_schema(config)
 
     # combiner without type
     del config["combiner"]["type"]
     with pytest.raises(ConfigValidationError):
-        ModelConfig.from_dict(config)
+        check_schema(config)
 
     # bad combiner type
     config["combiner"]["type"] = "fake"
     with pytest.raises(ConfigValidationError):
-        ModelConfig.from_dict(config)
+        check_schema(config)
 
     # bad combiner format (list instead of dict)
     config["combiner"] = [{"type": "tabnet"}]
     with pytest.raises(ConfigValidationError):
-        ModelConfig.from_dict(config)
+        check_schema(config)
 
     # bad combiner parameter types
     config["combiner"] = {
@@ -87,7 +88,7 @@ def test_config_bad_combiner():
         "dropout": False,
     }
     with pytest.raises(ConfigValidationError):
-        ModelConfig.from_dict(config)
+        check_schema(config)
 
     # bad combiner parameter range
     config["combiner"] = {
@@ -95,7 +96,7 @@ def test_config_bad_combiner():
         "dropout": -1,
     }
     with pytest.raises(ConfigValidationError):
-        ModelConfig.from_dict(config)
+        check_schema(config)
 
 
 def test_config_bad_combiner_types_enums():
@@ -109,66 +110,66 @@ def test_config_bad_combiner_types_enums():
     }
 
     # config is valid at this point
-    ModelConfig.from_dict(config)
+    check_schema(config)
 
     # Test weights initializer:
     config["combiner"]["weights_initializer"] = {"test": "fail"}
     with pytest.raises(ConfigValidationError):
-        ModelConfig.from_dict(config)
+        check_schema(config)
     config["combiner"]["weights_initializer"] = "fail"
     with pytest.raises(ConfigValidationError):
-        ModelConfig.from_dict(config)
+        check_schema(config)
     config["combiner"]["weights_initializer"] = {}
     with pytest.raises(ConfigValidationError):
-        ModelConfig.from_dict(config)
+        check_schema(config)
     config["combiner"]["weights_initializer"] = {"type": "fail"}
     with pytest.raises(ConfigValidationError):
-        ModelConfig.from_dict(config)
+        check_schema(config)
     config["combiner"]["weights_initializer"] = {"type": "normal", "stddev": 0}
-    ModelConfig.from_dict(config)
+    check_schema(config)
 
     # Test bias initializer:
     del config["combiner"]["weights_initializer"]
     config["combiner"]["bias_initializer"] = "kaiming_uniform"
-    ModelConfig.from_dict(config)
+    check_schema(config)
     config["combiner"]["bias_initializer"] = "fail"
     with pytest.raises(ConfigValidationError):
-        ModelConfig.from_dict(config)
+        check_schema(config)
     config["combiner"]["bias_initializer"] = {}
     with pytest.raises(ConfigValidationError):
-        ModelConfig.from_dict(config)
+        check_schema(config)
     config["combiner"]["bias_initializer"] = {"type": "fail"}
     with pytest.raises(ConfigValidationError):
-        ModelConfig.from_dict(config)
+        check_schema(config)
     config["combiner"]["bias_initializer"] = {"type": "zeros", "stddev": 0}
-    ModelConfig.from_dict(config)
+    check_schema(config)
 
     # Test norm:
     del config["combiner"]["bias_initializer"]
     config["combiner"]["norm"] = "batch"
-    ModelConfig.from_dict(config)
+    check_schema(config)
     config["combiner"]["norm"] = "fail"
     with pytest.raises(ConfigValidationError):
-        ModelConfig.from_dict(config)
+        check_schema(config)
 
     # Test activation:
     del config["combiner"]["norm"]
     config["combiner"]["activation"] = "relu"
-    ModelConfig.from_dict(config)
+    check_schema(config)
     config["combiner"]["activation"] = 123
     with pytest.raises(ConfigValidationError):
-        ModelConfig.from_dict(config)
+        check_schema(config)
 
     # Test reduce_output:
     del config["combiner"]["activation"]
     config2 = {**config}
     config2["combiner"]["type"] = "tabtransformer"
     config2["combiner"]["reduce_output"] = "sum"
-    ModelConfig.from_dict(config)
+    check_schema(config)
     config2["combiner"]["reduce_output"] = "fail"
     with pytest.raises(ConfigValidationError):
-        ModelConfig.from_dict(config2)
+        check_schema(config2)
 
     # Test reduce_output = None:
     config2["combiner"]["reduce_output"] = None
-    ModelConfig.from_dict(config2)
+    check_schema(config2)
