+++ tests/ludwig/config_validation/test_validate_config_preprocessing.py
@@ -1,7 +1,4 @@
-import pytest
-
-from ludwig.error import ConfigValidationError
-from ludwig.schema.model_types.base import ModelConfig
+from ludwig.config_validation.validation import check_schema
 from tests.integration_tests.utils import binary_feature, category_feature
 
 
@@ -21,43 +18,10 @@ def test_config_preprocessing():
         },
     }
 
-    ModelConfig.from_dict(config)
+    check_schema(config)
 
     # TODO(ksbrar): Circle back after discussing whether additional properties should be allowed long-term.
     # config["preprocessing"]["fake_parameter"] = True
 
     # with pytest.raises(Exception):
     #     ModelConfig(config)
-
-
-def test_balance_non_binary_failure():
-    config = {
-        "input_features": [
-            {"name": "Index", "proc_column": "Index", "type": "number"},
-            {"name": "random_1", "proc_column": "random_1", "type": "number"},
-            {"name": "random_2", "proc_column": "random_2", "type": "number"},
-        ],
-        "output_features": [{"name": "Label", "proc_column": "Label", "type": "number"}],
-        "preprocessing": {"oversample_minority": 0.2},
-    }
-
-    with pytest.raises(ConfigValidationError):
-        ModelConfig.from_dict(config)
-
-
-def test_balance_multiple_class_failure():
-    config = {
-        "input_features": [
-            {"name": "Index", "proc_column": "Index", "type": "number"},
-            {"name": "random_1", "proc_column": "random_1", "type": "number"},
-            {"name": "random_2", "proc_column": "random_2", "type": "number"},
-        ],
-        "output_features": [
-            {"name": "Label", "proc_column": "Label", "type": "binary"},
-            {"name": "Label2", "proc_column": "Label2", "type": "binary"},
-        ],
-        "preprocessing": {"oversample_minority": 0.2},
-    }
-
-    with pytest.raises(ConfigValidationError):
-        ModelConfig.from_dict(config)
