+++ tests/ludwig/config_validation/test_validate_config_trainer.py
@@ -1,8 +1,8 @@
 import pytest
 
+from ludwig.config_validation.validation import check_schema
 from ludwig.constants import TRAINER
 from ludwig.error import ConfigValidationError
-from ludwig.schema.model_types.base import ModelConfig
 from ludwig.schema.optimizers import optimizer_registry
 from ludwig.schema.trainer import ECDTrainerConfig
 from tests.integration_tests.utils import binary_feature, category_feature, number_feature
@@ -23,14 +23,14 @@ def test_config_trainer_empty_null_and_default():
         },
         TRAINER: {},
     }
-    ModelConfig.from_dict(config)
+    check_schema(config)
 
     config[TRAINER] = None
     with pytest.raises(ConfigValidationError):
-        ModelConfig.from_dict(config)
+        check_schema(config)
 
     config[TRAINER] = ECDTrainerConfig.Schema().dump({})
-    ModelConfig.from_dict(config)
+    check_schema(config)
 
 
 def test_config_trainer_bad_optimizer():
@@ -45,26 +45,26 @@ def test_config_trainer_bad_optimizer():
         },
         TRAINER: {},
     }
-    ModelConfig.from_dict(config)
+    check_schema(config)
 
     # Test manually set-to-null optimizer vs unspecified:
     config[TRAINER]["optimizer"] = None
     with pytest.raises(ConfigValidationError):
-        ModelConfig.from_dict(config)
+        check_schema(config)
     assert ECDTrainerConfig.Schema().load({}).optimizer is not None
 
     # Test all types in optimizer_registry supported:
     for key in optimizer_registry.keys():
         config[TRAINER]["optimizer"] = {"type": key}
-        ModelConfig.from_dict(config)
+        check_schema(config)
 
     # Test invalid optimizer type:
     config[TRAINER]["optimizer"] = {"type": 0}
     with pytest.raises(ConfigValidationError):
-        ModelConfig.from_dict(config)
+        check_schema(config)
     config[TRAINER]["optimizer"] = {"type": "invalid"}
     with pytest.raises(ConfigValidationError):
-        ModelConfig.from_dict(config)
+        check_schema(config)
 
 
 def test_optimizer_property_validation():
@@ -79,30 +79,30 @@ def test_optimizer_property_validation():
         },
         TRAINER: {},
     }
-    ModelConfig.from_dict(config)
+    check_schema(config)
 
     # Test that an optimizer's property types are enforced:
     config[TRAINER]["optimizer"] = {"type": "rmsprop"}
-    ModelConfig.from_dict(config)
+    check_schema(config)
 
     config[TRAINER]["optimizer"]["momentum"] = "invalid"
     with pytest.raises(ConfigValidationError):
-        ModelConfig.from_dict(config)
+        check_schema(config)
 
     # Test extra keys are excluded and defaults are loaded appropriately:
     config[TRAINER]["optimizer"]["momentum"] = 10
     config[TRAINER]["optimizer"]["extra_key"] = "invalid"
-    ModelConfig.from_dict(config)
+    check_schema(config)
     assert not hasattr(ECDTrainerConfig.Schema().load(config[TRAINER]).optimizer, "extra_key")
 
     # Test bad parameter range:
     config[TRAINER]["optimizer"] = {"type": "rmsprop", "eps": -1}
     with pytest.raises(ConfigValidationError):
-        ModelConfig.from_dict(config)
+        check_schema(config)
 
     # Test config validation for tuple types:
     config[TRAINER]["optimizer"] = {"type": "adam", "betas": (0.1, 0.1)}
-    ModelConfig.from_dict(config)
+    check_schema(config)
 
 
 def test_clipper_property_validation():
@@ -117,13 +117,13 @@ def test_clipper_property_validation():
         },
         TRAINER: {},
     }
-    ModelConfig.from_dict(config)
+    check_schema(config)
 
     # Test null/empty clipper:
     config[TRAINER]["gradient_clipping"] = None
-    ModelConfig.from_dict(config)
+    check_schema(config)
     config[TRAINER]["gradient_clipping"] = {}
-    ModelConfig.from_dict(config)
+    check_schema(config)
     assert (
         ECDTrainerConfig.Schema().load(config[TRAINER]).gradient_clipping
         == ECDTrainerConfig.Schema().load({}).gradient_clipping
@@ -132,19 +132,19 @@ def test_clipper_property_validation():
     # Test invalid clipper type:
     config[TRAINER]["gradient_clipping"] = 0
     with pytest.raises(ConfigValidationError):
-        ModelConfig.from_dict(config)
+        check_schema(config)
     config[TRAINER]["gradient_clipping"] = "invalid"
     with pytest.raises(ConfigValidationError):
-        ModelConfig.from_dict(config)
+        check_schema(config)
 
     # Test that an optimizer's property types are enforced:
     config[TRAINER]["gradient_clipping"] = {"clipglobalnorm": None}
-    ModelConfig.from_dict(config)
+    check_schema(config)
     config[TRAINER]["gradient_clipping"] = {"clipglobalnorm": 1}
-    ModelConfig.from_dict(config)
+    check_schema(config)
     config[TRAINER]["gradient_clipping"] = {"clipglobalnorm": "invalid"}
     with pytest.raises(ConfigValidationError):
-        ModelConfig.from_dict(config)
+        check_schema(config)
 
     # Test extra keys are excluded and defaults are loaded appropriately:
     config[TRAINER]["gradient_clipping"] = {"clipnorm": 1}
